refactor: ShopCategory id Long -> Integer 수정

+++ src/main/java/in/koreatech/koin/domain/shop/dto/ShopCategoriesResponse.java
@@ -17,23 +17,26 @@ public record ShopCategoriesResponse(
     @Schema(description = "모든 상점 카테고리 리스트")
     List<InnerShopCategory> shopCategories
 ) {
+
     public static ShopCategoriesResponse from(List<ShopCategory> shopCategories) {
         return new ShopCategoriesResponse(
             shopCategories.size(),
             shopCategories.stream().map(InnerShopCategory::from).toList()
         );
     }
+
     @JsonNaming(value = SnakeCaseStrategy.class)
-    private record InnerShopCategory (
+    private record InnerShopCategory(
         @Schema(example = "2", description = "고유 id")
-        Long id,
+        Integer id,
 
         @Schema(example = "https://static.koreatech.in/test.png", description = "이미지 URL")
         String imageUrl,
 
         @Schema(example = "치킨", description = "이름")
         String name
-    ){
+    ) {
+
         public static InnerShopCategory from(ShopCategory shopCategory) {
             return new InnerShopCategory(
                 shopCategory.getId(),
